+++ nv/vi/settings.py
@@ -212,6 +212,30 @@ def set_cmdline_cwd(path):
     _storage['cmdline_cwd'] = path
 
 
+def _toggle_preference(name):
+    preferences = load_settings('Preferences.sublime-settings')
+    value = preferences.get(name)
+    preferences.set(name, not value)
+    save_settings('Preferences.sublime-settings')
+
+
+def toggle_ctrl_keys():
+    _toggle_preference('vintageous_use_ctrl_keys')
+
+
+def toggle_super_keys():
+    _toggle_preference('vintageous_use_super_keys')
+
+
+def toggle_side_bar(window):
+    window.run_command('toggle_side_bar')
+
+    if window.is_sidebar_visible():
+        window.run_command('focus_side_bar')
+    else:
+        window.focus_group(window.active_group())
+
+
 def _get_option(view, name):
     name = _resolve_option_alias(name)